Build: v2.0.4
Added: New errors, checks if queue already contains a loader with same id. Checks for circular reference e.g. when adding an IAssetLoader into itself or when an IAssetLoader is part of it own loading queue, even if nested within other loaders. Errors if a loader is added to more than one parent at the same time.

Added: URLParser, a class that handles all urls, primary function is to see if a string is a valid url or not.

Added: Test cases for all the above added.